+++ pylinac/ct.py
@@ -2673,11 +2673,20 @@ class CatPhanBase(ResultsDataMixin[CatphanResult], QuaacMixin):
             except Exception:
                 pass
 
-    def plotly_side_view(self, show_legend: bool) -> go.Figure:
+    def plotly_side_view(self, show_legend: bool, **kwargs) -> go.Figure:
+        """Plot a view of the scan from the side with lines showing detected module positions.
+
+        Parameters
+        ----------
+        show_legend: bool
+            Whether to show the plot legend.
+        kwargs
+            Arguments passed to the axis constructor.
+        """
         fig = go.Figure()
         side_array = self.dicom_stack.side_view(axis=1)
         add_title(fig, "Side View")
-        fig.add_heatmap(z=side_array, colorscale="gray", showscale=False)
+        fig.add_heatmap(z=side_array, colorscale="gray", showscale=False, **kwargs)
 
         for module in self._detected_modules():
             add_vertical_line(
@@ -2690,12 +2699,22 @@ class CatPhanBase(ResultsDataMixin[CatphanResult], QuaacMixin):
         fig.update_layout(showlegend=show_legend)
         return fig
 
-    def plot_side_view(self, axis: Axes) -> None:
-        """Plot a view of the scan from the side with lines showing detected module positions"""
+    def plot_side_view(self, axis: Axes, **kwargs) -> None:
+        """Plot a view of the scan from the side with lines showing detected module positions.
+
+        Parameters
+        ----------
+        axis: Axes
+            The axis to plot the scan to
+        kwargs
+            Arguments passed to the axis constructor.
+        """
         side_array = self.dicom_stack.side_view(axis=1)
         axis.set_yticks([])
         axis.set_title("Side View")
-        axis.imshow(side_array, aspect="auto", cmap="gray", interpolation="none")
+        axis.imshow(
+            side_array, aspect="auto", cmap="gray", interpolation="none", **kwargs
+        )
         for module in self._detected_modules():
             axis.axvline(module.slice_num)